cropsense/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
tensorflow>=2.16.1; sys_platform != 'win32'
tensorflow-intel>=2.16.1; sys_platform == 'win32' # Better for Windows
scikit-learn>=1.5.0
xgboost>=2.0.3
pandas>=2.2.2
numpy>=1.26.4
Pillow>=10.3.0
joblib>=1.4.2
python-multipart>=0.0.9
pydantic>=2.7.1
matplotlib>=3.9.0
